Ignore missing Javadoc comments or tags
JDK-15 introduced a change for missing Javadocs: https://bugs.openjdk.java.net/browse/JDK-8242607.
That produces a lot of warnings like the following one:
```
/vividus/vividus/src/main/java/org/vividus/configuration/Vividus.java:35: warning: no comment
public final class Vividus
```
It doesn't break builds, but bring a lot of noise.
This change addresses this issue and hides the warnings (bear in mind: it also hides warnings for missing Javadoc tags, so in case of issue the better solution might be required)