Do not enable Java Security Manager
Java 24 does not support Security Manager anymore: A command line option has attempted to allow or enable the Security Manager. Enabling a Security Manager is not supported.
At the same time Kafka team has implemnted enhancements allowing to run Kafka client on Java 24, so the workaround is not needed anymore, but user must use VIVIDUS 0.6.15 (which includes kafka-clients 3.9.1) or higher.